+++ backend/src/index.ts
@@ -22,10 +22,6 @@ const PORT: number = parseInt(process.env.PORT as string, 10);
 
 const app = express();
 
-app.get('/', (req, res) => {
-    res.send('Hello World!!!')
-})
-
 // App Configuration
 
 // Enables and sets defaults for a myriad of small middleware packages that set HTTP response headers. It's VERY important:
@@ -34,6 +30,8 @@ app.use(helmet());
 app.use(cors());
 // Stands in for the body-parser used in BlockLand; parses incoming requests into JSON format and enables you to access the request BODY:
 app.use(express.json());
+// Set the public directory as the server's static 
+app.use(express.static("public"));
 
 // Server Endpoints
 app.get('/api/:dbName/:collection', getTestData)