+++ frontend/sketch.js
@@ -1,7 +1,0 @@
-function setup () {
-    createCanvas(512, 256);
-}
-
-function draw() {
-    background(0);
-}